% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnBreak.R
\name{addColumnBreak}
\alias{addColumnBreak}
\alias{addColumnBreak.docx}
\title{Add a column break into a section}
\usage{
addColumnBreak(doc, ...)

\method{addColumnBreak}{docx}(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a column break into a section
}
\details{
\code{addColumnBreak} only works with docx documents.
}
\examples{
\donttest{
doc.filename = "add_col_break.docx"
doc = docx( )
doc = addSection(doc, ncol = 2, columns.only = TRUE )
doc = addParagraph( doc = doc, "Text 1.", "Normal" )
doc = addColumnBreak(doc )
doc = addParagraph( doc = doc, "Text 2.", "Normal" )
writeDoc( doc, file = doc.filename )
}
}
\seealso{
\code{\link{docx}}, \code{\link{addSection}}
}

