% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addFooter}
\alias{addFooter}
\title{Insert a footer into a document object}
\usage{
addFooter(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Insert a footer into a document object
}
\details{
addFooter only works for pptx documents.
}
\seealso{
\code{\link{pptx}}
}

