% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIframe.bsdoc.R
\name{addIframe.bsdoc}
\alias{addIframe.bsdoc}
\title{Insert an iframe into a bsdoc object}
\usage{
\method{addIframe}{bsdoc}(doc, src, width, height, seamless = FALSE,
  par.properties = parProperties(text.align = "center", padding = 5), ...)
}
\arguments{
\item{doc}{\code{\link{bsdoc}} object where iframe has to be added}

\item{src}{url of the document to embed in the iframe}

\item{width}{Specifies the width of an iframe}

\item{height}{Specifies the height of an iframe}

\item{seamless}{Specifies that the iframe should look like it is a part of the containing document}

\item{par.properties}{paragraph formatting properties of the paragraph that contains iframe. 
An object of class \code{\link{parProperties}}}

\item{...}{further arguments, not used.}
}
\value{
an object of class \code{\link{bsdoc}}.
}

