% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addIframe}
\alias{addIframe}
\title{Add an iframe into a document object}
\usage{
addIframe(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add an iframe into a document object
}
\seealso{
\code{\link{addIframe.bsdoc}}
}

