% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImage.bsdoc.R
\name{addImage.bsdoc}
\alias{addImage.bsdoc}
\title{Insert an external image into a bsdoc object}
\usage{
\method{addImage}{bsdoc}(doc, filename, width, height,
  par.properties = parProperties(text.align = "center", padding = 5), ...)
}
\arguments{
\item{doc}{\code{\link{bsdoc}} object where external image has to be added}

\item{filename}{\code{"character"} value, complete filename of the external image}

\item{width}{image width in inches}

\item{height}{image height in inches}

\item{par.properties}{paragraph formatting properties of the paragraph that contains images.
An object of class \code{\link{parProperties}}}

\item{...}{further arguments, not used.}
}
\value{
an object of class \code{\link{bsdoc}}.
}
\description{
Add an external image into a \code{\link{bsdoc}} object.
}
\details{
Arguments \code{width} and \code{height} can be defined with functions
\code{png::readPNG}, \code{jpeg::readJPEG} or \code{bmp::read.bmp}.
}
\examples{
#START_TAG_TEST
doc.filename = "addImage_bsdoc/example.html"

# set default font size to 11
options( "ReporteRs-fontsize" = 11 )

doc = bsdoc( )
# files 'logo.jpg'
img.file1 = file.path( Sys.getenv("R_HOME"), "doc", "html", "logo.jpg" )

if( file.exists( img.file1 ) && requireNamespace("jpeg", quietly = TRUE) ){
  dims <- attr( jpeg::readJPEG(img.file1), "dim" )
  doc = addTitle( doc, "Add images with width and height", level = 1)
  doc = addImage(doc, img.file1, width = dims[2]/72, height = dims[1]/72)
}

# Write the object
writeDoc( doc, file = doc.filename )
#STOP_TAG_TEST
}
\seealso{
\code{\link{bsdoc}}, \code{\link{addPlot.bsdoc}}
, \code{\link{addImage}}
}

