\name{addMarkdown}
\alias{addMarkdown}
\title{Add a markdown text or file}
\usage{
addMarkdown(doc, file, text, ...)
}
\arguments{
  \item{doc}{document object}

  \item{file}{markdown file. Not used if text is provided.}

  \item{text}{character vector. The markdown to parse.}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add markdown into a document object
}
\seealso{
\code{\link{docx}}, \code{\link{addMarkdown.docx}}
}

