% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addPageBreak}
\alias{addPageBreak}
\title{Add a page break into a document object}
\usage{
addPageBreak(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a page break into a document object
}
\details{
\code{addPageBreak} only works with docx documents.

See \code{\link{addPageBreak.docx}} for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{addPageBreak.docx}}
}

