% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Footnote.R
\name{addParagraph.Footnote}
\alias{addParagraph.Footnote}
\title{Insert a paragraph into a Footnote object}
\usage{
\method{addParagraph}{Footnote}(doc, value, par.properties = parProperties(),
  ...)
}
\arguments{
\item{doc}{\code{\link{Footnote}} object where to add paragraphs.}

\item{value}{text to add to the document as paragraphs:
an object of class \code{\link{pot}} or \code{\link{set_of_paragraphs}}
or a character vector.}

\item{par.properties}{\code{\link{parProperties}} to apply to paragraphs.}

\item{...}{further arguments, not used.}
}
\value{
an object of class \code{\link{Footnote}}.
}
\description{
Insert paragraph(s) of text into a \code{Footnote}. To create a \code{\link{Footnote}} made of
several paragraphs with different \code{\link{parProperties}}, add sequentially
paragraphs with their associated \code{parProperties} objects with this function.
}
\seealso{
\code{\link{Footnote}}, \code{\link{parProperties}}, \code{\link{pot}}
, \code{\link{set_of_paragraphs}}
}

