% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlot.bsdoc.R
\name{addPlot.bsdoc}
\alias{addPlot.bsdoc}
\title{Add a plot into an bsdoc object}
\usage{
\method{addPlot}{bsdoc}(doc, fun, pointsize = getOption("ReporteRs-fontsize"),
  vector.graphic = T, width = 6, height = 6,
  fontname = getOption("ReporteRs-default-font"),
  fontname_serif = "Times New Roman", fontname_sans = "Calibri",
  fontname_mono = "Courier New", fontname_symbol = "Symbol",
  par.properties = parCenter(padding = 5), ...)
}
\arguments{
\item{doc}{Object of class \code{bsdoc} where paragraph has to be added}

\item{fun}{plot function. The function will be executed to produce graphics.
For \code{grid} or \code{lattice} or \code{ggplot} object, the function
should just be print and an extra argument x should specify the object
to plot. For traditionnal plots, the function should contain plot instructions. See examples.}

\item{pointsize}{the default pointsize of plotted text in pixels, default to 12.}

\item{vector.graphic}{logical scalar, default to FALSE. If TRUE, vector graphics
are produced instead of PNG images. If TRUE, SVG is produced.}

\item{width}{plot width in inches (default value is 6).}

\item{height}{plot height in inches (default value is 6).}

\item{fontname}{deprecated. the default font family to use, default to getOption("ReporteRs-default-font").}

\item{fontname_serif, fontname_sans, fontname_mono, fontname_symbol}{font
names for font faces.
Used fonts should be available in the operating system.}

\item{par.properties}{paragraph formatting properties of the paragraph that contains plot(s). An object of class \code{\link{parProperties}}}

\item{...}{arguments for \code{fun}.}
}
\value{
an object of class \code{\link{bsdoc}}.
}
\description{
Add a plot into the \code{bsdoc} object.
}
\examples{
#START_TAG_TEST
doc.filename = "addPlot_bsdoc/example.html"

# set default font size to 11
options( "ReporteRs-fontsize" = 11 )

doc = bsdoc( )

doc = addTitle( doc, "Title example 1", level = 1 )
# Add a base plot
# set vector.graphic to FALSE if Word version
#   used to read the file is <= 2007
doc = addPlot( doc, fun = plot,
  x = rnorm( 100 ), y = rnorm (100 ), main = "base plot main title",
  vector.graphic = TRUE,
  width = 5, height = 7,
  par.properties = parProperties(text.align = "left")
  )

doc = addTitle( doc, "Title example 2", level = 1 )
# load ggplot2
if( requireNamespace("ggplot2", quietly = TRUE) ){

  # create a ggplot2 plot
  myplot = ggplot2::qplot(Sepal.Length, Petal.Length, data = iris
    , color = Species, size = Petal.Width, alpha = I(0.7) )

  # Add myplot into object doc
  #   myplot is assigned to argument 'x' because function 'print' on ggplot
  #   objects is expecting argument 'x'.
  doc = addPlot( doc = doc, fun = print, x = myplot )
}

writeDoc( doc, file = doc.filename )
#STOP_TAG_TEST
}
\seealso{
\code{\link{bsdoc}}, \code{\link{addPlot}}
}

