\name{addSlide}
\alias{addSlide}
\title{Add a slide into a document object}
\usage{
  addSlide(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
  a document object
}
\description{
  Add a slide into a document object
}
\seealso{
  \code{\link{pptx}}, \code{\link{addSlide.pptx}}
}

