% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addSubtitle}
\alias{addSubtitle}
\title{Add a subtitle shape into a document object}
\usage{
addSubtitle(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a subtitle shape into a document object
}
\details{
\code{addSubtitle} only works with pptx documents. See \code{\link{addSubtitle.pptx}} for examples.
}
\seealso{
\code{\link{pptx}}, \code{\link{addSubtitle.pptx}}
}

