\name{addTitle}
\alias{addTitle}
\title{Add a title into a document object}
\usage{
  addTitle(doc, value, ...)
}
\arguments{
  \item{doc}{document object}

  \item{value}{\code{"character"} value to use as title
  text}

  \item{...}{further arguments passed to or from other
  methods..}
}
\value{
  a document object
}
\description{
  Add a title into a document object
}
\seealso{
  \code{\link{docx}}, \code{\link{addTitle.docx}},
  \code{\link{pptx}} , \code{\link{addTitle.pptx}}
}

