% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addTOC}
\alias{addTOC}
\title{Add a table of contents into a document object}
\usage{
addTOC(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a table of contents into a document object
}
\details{
\code{addTOC} only works with docx documents.

See \code{\link{addTOC.docx}} for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{addTOC.docx}}, \code{\link{styles.docx}}
}

