% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{as.FlexTable}
\alias{as.FlexTable}
\title{R tables as FlexTables}
\usage{
as.FlexTable(x, ...)
}
\arguments{
\item{x}{object to get \code{FlexTable} from}

\item{...}{further arguments passed to other methods}
}
\value{
a \code{\link{FlexTable}} object
}
\description{
Get a \code{\link{FlexTable}} object from
an R object.
}
\seealso{
\code{\link{FlexTable}}
}

