% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FlexTable.sessionInfo.R
\name{as.FlexTable.sessionInfo}
\alias{as.FlexTable.sessionInfo}
\title{get FlexTable from a sessionInfo object}
\usage{
\method{as.FlexTable}{sessionInfo}(x, locale = TRUE, ...)
}
\arguments{
\item{x}{\code{sessionInfo} object to get \code{FlexTable} from}

\item{locale}{show locale information?}

\item{...}{further arguments, not used.}
}
\value{
a \code{\link{FlexTable}} object
}
\description{
Get a \code{\link{FlexTable}} object from 
a \code{\link{sessionInfo}} object.
}
\examples{
#
as.FlexTable( sessionInfo() )
}

