% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDoc.R
\name{as.html.bsdoc}
\alias{as.html.bsdoc}
\title{get HTML code from a bsdoc object}
\usage{
\method{as.html}{bsdoc}(object, ...)
}
\arguments{
\item{object}{the \code{bsdoc} object}

\item{...}{further arguments passed to other methods}
}
\value{
a character value
}
\description{
get HTML code from a bsdoc document
}
\seealso{
\code{\link{bsdoc}}
}

