% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.html.FlexTable.R
\name{as.html.FlexTable}
\alias{as.html.FlexTable}
\title{get HTML code from a FlexTable}
\usage{
\method{as.html}{FlexTable}(object, ...)
}
\arguments{
\item{object}{the \code{FlexTable} object}

\item{...}{further arguments passed to other methods}
}
\value{
a character value
}
\description{
get HTML code from a FlexTable
}
\seealso{
\code{\link{FlexTable}}
}
