% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{as.html}
\alias{as.html}
\title{get HTML code}
\usage{
as.html(object, ...)
}
\arguments{
\item{object}{object to get HTML from}

\item{...}{further arguments passed to other methods}
}
\value{
a character value
}
\description{
Get HTML code in a character vector.
}
\details{
See \code{\link{FlexTable}} or \code{\link{raphael.html}} for examples.
}
\seealso{
\code{\link{FlexTable}}, \code{\link{raphael.html}}
}

