\name{borderDashed}
\alias{borderDashed}
\title{shortcut for dashed border}
\usage{
borderDashed(...)
}
\arguments{
  \item{...}{arguments passed to borderProperties}
}
\description{
shortcut for a dashed border borderProperties()
}
\examples{
borderDashed()
}
\seealso{
\code{\link{textNormal}} , \code{\link{textBold}} ,
\code{\link{textItalic}} , \code{\link{textBoldItalic}} ,
\code{\link{parRight}} , \code{\link{parLeft}} ,
\code{\link{parCenter}} , \code{\link{parJustify}} ,
\code{\link{borderDotted}} , \code{\link{borderNone}} ,
\code{\link{borderSolid}}
}

