% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/format.properties.R
\name{borderNone}
\alias{borderNone}
\title{shortcut for no border}
\usage{
borderNone(...)
}
\arguments{
\item{...}{arguments passed to borderProperties}
}
\description{
shortcut for no border borderProperties()
}
\examples{
borderNone()
}
\seealso{
\code{\link{textNormal}}
, \code{\link{textBold}}
, \code{\link{textItalic}}
, \code{\link{textBoldItalic}}
, \code{\link{parRight}}
, \code{\link{parLeft}}
, \code{\link{parCenter}}
, \code{\link{parJustify}}
, \code{\link{borderDotted}}
, \code{\link{borderDashed}}
, \code{\link{borderSolid}}
}

