% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tools.R
\name{check_valid_java_version}
\alias{check_valid_java_version}
\title{check validity of java version for the package}
\usage{
check_valid_java_version()
}
\description{
check that the java version is <= 1.8
}
