% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{chprop}
\alias{chprop}
\title{Change a formatting properties object}
\usage{
chprop(object, ...)
}
\arguments{
\item{object}{formatting properties object}

\item{...}{further arguments passed to other methods}
}
\value{
a formatting properties object
}
\description{
Change a formatting properties object
}
\details{
See \code{\link{chprop.textProperties}} or \code{\link{chprop.parProperties}}
or \code{\link{chprop.cellProperties}} for examples.
}
\seealso{
\code{\link{cellProperties}}, \code{\link{textProperties}}
, \code{\link{parProperties}}
}
