% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CodeBlock.R
\name{CodeBlock}
\alias{CodeBlock}
\title{Code Block Object}
\usage{
CodeBlock(file, text, text.properties = textProperties(color = "#A7947D"),
  par.properties = parProperties(text.align = "left", shading.color =
  "#5FB0B8"))
}
\arguments{
\item{file}{script file. Not used if text is provided.}

\item{text}{character vector. The text to parse. Not used if file is provided.}

\item{text.properties}{default textProperties object}

\item{par.properties}{default parProperties object}
}
\description{
Code Block Object. A code block object is a block of text
treated as verbatim text in a document object.
}
\examples{
#START_TAG_TEST
cb_example = CodeBlock( text = "ls -a\\nwhich -a ls" )
#STOP_TAG_TEST
}
\seealso{
\code{\link{addCodeBlock}}
}

