% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorProperties.R
\name{colorProperties}
\alias{colorProperties}
\alias{print.colorProperties}
\alias{as.character.colorProperties}
\title{color properties object}
\usage{
colorProperties(value = "black")

\method{print}{colorProperties}(x, ...)

\method{as.character}{colorProperties}(x, ...)
}
\arguments{
\item{value}{valid color}

\item{x}{a colorProperties object}

\item{...}{unused}
}
\description{
create a color properties object.
}

