\docType{data}
\name{data_ReporteRs}
\alias{data_ReporteRs}
\title{Dummy dataset used in ReporterRs examples}
\format{A data frame with 4 rows and 6 variables}
\description{
  A dataset containing several data types and few rows
}
\details{
  \itemize{ \item col1 a character vector \item col2 an
  integer vector \item col3 a double vector \item col4 a
  date vector \item col5 a double vector (percent values)
  \item col6 a boolean vector }
}
\keyword{datasets}

