% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{declareTitlesStyles}
\alias{declareTitlesStyles}
\title{Set manually headers'styles of a document object}
\usage{
declareTitlesStyles(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Set manually titles'styles of a document object
}
\details{
\code{declareTitlesStyles} only works with docx documents.

See \code{\link{declareTitlesStyles.docx}} for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{styles.docx}}
, \code{\link{declareTitlesStyles.docx}}, \code{\link{addTOC.docx}}
}

