\name{deleteBookmarkNextContent}
\alias{deleteBookmarkNextContent}
\title{delete first content after a bookmark into a docx object}
\usage{
  deleteBookmarkNextContent(doc, bookmark)
}
\arguments{
  \item{doc}{Object of class \code{"docx"}}

  \item{bookmark}{a character vector specifying bookmark id
  to delete}
}
\description{
  delete first content after a bookmark into a docx object
}
\seealso{
  \code{\link{docx}}
}

