\name{FlexHeader[<-}
\alias{FlexHeader[<-}
\title{modify FlexTable header}
\usage{
FlexHeader[(x, i, j) <- value
}
\arguments{
  \item{x}{sqdqd}

  \item{i}{sdfdf}

  \item{j}{sdfd sdf}

  \item{value}{value sd}
}
\description{
Add and modify content or formatting properties to
FlexTable header rows
}

