% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexRow.R
\name{[<-.FlexRow}
\alias{[<-.FlexRow}
\alias{alterFlexRow}
\title{modify FlexRow content}
\usage{
\method{[}{FlexRow}(x, i) <- value
}
\arguments{
\item{x}{the \code{FlexRow} object}

\item{i}{a single integer value.}

\item{value}{an object of class \code{\link{FlexCell}}}
}
\description{
add or replace FlexCell into a FlexRow object
}
\seealso{
\code{\link{FlexTable}}, \code{\link{addFlexTable}}, \code{\link{FlexRow}}
, \code{\link{addHeaderRow}}, \code{\link{addFooterRow}}
}

