% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FontMetric.R
\name{FontMetric}
\alias{FontMetric}
\title{Font metric}
\usage{
FontMetric(fontfamily, fontsize)
}
\arguments{
\item{fontfamily}{font name}

\item{fontsize}{font size}
}
\description{
get font metric from a font name and a size
}

