\name{get.darker.tableProperties}
\alias{get.darker.tableProperties}
\title{get a darker tableProperties}
\usage{
get.darker.tableProperties()
}
\description{
darker tableProperties object
}
\examples{
get.darker.tableProperties()
}
\seealso{
\code{\link{addTable}} ,
\code{\link{get.default.tableProperties}} ,
\code{\link{get.light.tableProperties}} ,
\code{\link{get.greenheader.tableProperties}}
}

