\name{get.default.tableProperties}
\alias{get.default.tableProperties}
\title{get default tableProperties}
\usage{
  get.default.tableProperties()
}
\description{
  default tableProperties object
}
\examples{
get.default.tableProperties()
}
\seealso{
  \code{\link{addTable}} ,
  \code{\link{get.light.tableProperties}} ,
  \code{\link{get.greenheader.tableProperties}} ,
  \code{\link{get.darker.tableProperties}}
}

