\name{get.greenheader.tableProperties}
\alias{get.greenheader.tableProperties}
\title{get a green header tableProperties}
\usage{
  get.greenheader.tableProperties()
}
\description{
  green header tableProperties object
}
\examples{
get.greenheader.tableProperties()
}
\seealso{
  \code{\link{addTable}} ,
  \code{\link{get.default.tableProperties}} ,
  \code{\link{get.light.tableProperties}} ,
  \code{\link{get.darker.tableProperties}}
}

