\name{get.light.tableProperties}
\alias{get.light.tableProperties}
\title{get a 'lighter' tableProperties}
\usage{
  get.light.tableProperties()
}
\description{
  light tableProperties object
}
\examples{
get.light.tableProperties()
}
\seealso{
  \code{\link{addTable}} ,
  \code{\link{get.default.tableProperties}} ,
  \code{\link{get.greenheader.tableProperties}} ,
  \code{\link{get.darker.tableProperties}}
}

