\docType{data}
\name{gforge_datacopy}
\alias{gforge_datacopy}
\title{melanoma dataset from boot package}
\format{A data frame with 4 rows and 10 variables}
\description{
  Transformed dataset to reproduce blog :
  http://gforge.se/2014/01/fast-track-publishing-using-knitr-part-iv/
}
\keyword{datasets}

