% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexTable.R
\name{knit_print.FlexTable}
\alias{knit_print.FlexTable}
\title{FlexTable custom printing function for knitr}
\usage{
\method{knit_print}{FlexTable}(x, ...)
}
\arguments{
\item{x}{a \code{FlexTable} to be printed}

\item{...}{further arguments, not used.}
}
\description{
FlexTable custom printing function for knitr
}
