% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot.R
\name{knit_print.pot}
\alias{knit_print.pot}
\title{pot custom printing function for knitr}
\usage{
\method{knit_print}{pot}(x, ...)
}
\arguments{
\item{x}{a \code{pot} to be printed}

\item{...}{further arguments, not used.}
}
\description{
pot custom printing function for knitr
}

