% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FlexTable_theme.R
\name{light.table}
\alias{light.table}
\title{get a simple FlexTable from a dataset}
\usage{
light.table(dataset, add.rownames = FALSE)
}
\arguments{
\item{dataset}{the data to use}

\item{add.rownames}{logical value - should the row.names be included in the table.}
}
\description{
get a simple FlexTable from a dataset
}
\examples{
#START_TAG_TEST
light.table( iris)
#STOP_TAG_TEST
}
\seealso{
\code{\link{FlexTable}}
}

