% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text_extract.R
\name{list_bookmarks}
\alias{list_bookmarks}
\title{List Bookmarks from a Word Document}
\usage{
list_bookmarks(x, body = TRUE, header = TRUE, footer = TRUE)
}
\arguments{
\item{x}{\code{\link{docx}} object}
}
\value{
a character vector
}
\description{
List all bookmarks available in a \code{docx} object.
}
\examples{
#START_TAG_TEST
doc = docx( title = "My example", template = file.path(
  find.package("ReporteRs"), "templates/bookmark_example.docx") )
list_bookmarks( doc )
#STOP_TAG_TEST
}
\seealso{
\code{\link{docx}}, \code{\link{text_extract}}
}

