% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_link.R
\name{office_web_viewer}
\alias{office_web_viewer}
\title{Office Web Viewer}
\usage{
office_web_viewer(url, width = "80\%", height = "500px")
}
\arguments{
\item{url}{file url}

\item{width}{iframe width - deprecated}

\item{height}{iframe height - deprecated}
}
\description{
Produce an iframe linked to Office Web Viewer. It let's you
display a Microsoft Office document in a Web browser.
}
