\name{ParagraphSection}
\alias{ParagraphSection}
\title{ParagraphSection (internal use only)}
\usage{
ParagraphSection(value, par.properties)
}
\arguments{
  \item{value}{a set_of_paragraphs object}

  \item{par.properties}{a parProperties object}
}
\description{
ParagraphSection (internal use only)
}

