% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pot.R
\name{pot_img}
\alias{pot_img}
\title{Image to be concatenate with pot object}
\usage{
pot_img(filename, width, height, ppi = 72)
}
\arguments{
\item{filename}{\code{"character"} value, complete filename of the external image}

\item{width}{image width in inches}

\item{height}{image height in inches}

\item{ppi}{dot per inches, default to 72}
}
\description{
Create an pot object that handle images.
}

