% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx-colors.R
\name{pptxColors}
\alias{pptxColors}
\alias{pptxColour}
\alias{pptxColors.pptx}
\alias{pptxColors.character}
\title{Colors in the powerpoint document}
\usage{
pptxColors(doc, full = FALSE, veryLight = 0.1, veryDark = 0.9)

pptxColour(doc, full = FALSE, veryLight = 0.1, veryDark = 0.9)

\method{pptxColors}{pptx}(doc, full = FALSE, veryLight = 0.1,
  veryDark = 0.9)

\method{pptxColors}{character}(doc, full = FALSE, veryLight = 0.1,
  veryDark = 0.9)
}
\arguments{
\item{doc}{Powerpoint object}

\item{full}{boolean to return full color matrix}

\item{veryLight}{a number betweeen 0 and 1 representing the cutoff
for very light colors.  This defaults to 0.1}

\item{veryDark}{a number between 0 and 1 representing the cutoff
for very dark colors.  This defaults to 0.9}
}
\value{
a list of named powerpoint based on the powerpoint theme.
}
\description{
This is based on the information about the powerpoint color scheme
given in the msdn article
https://msdn.microsoft.com/en-us/library/cc964302(v=office.12).aspx
The article is unclear what "very light" or "very dark" is, so it
is assumed the HSL lightness score is below 10% for very light
colors and the HSL lightness score is above 90% for very dark
colors. These values can be tuned.
}
\author{
Matthew L. Fidler
}
