% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsdoc.R
\name{print.bsdoc}
\alias{print.bsdoc}
\title{print informations about an object of class \code{\link{bsdoc}}.}
\usage{
\method{print}{bsdoc}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{bsdoc}}}

\item{...}{further arguments, not used.}
}
\description{
print informations about an object of class \code{\link{bsdoc}}.
}
\examples{
\dontrun{
# Create a new document
doc = bsdoc( )
print( doc )
}
}
\seealso{
\code{\link{bsdoc}}, \code{\link{print}}
}

