% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.docx.R
\name{print.docx}
\alias{print.docx}
\title{print informations about an object of class \code{\link{docx}}.}
\usage{
\method{print}{docx}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{docx}}}

\item{...}{further arguments, not used.}
}
\description{
print informations about an object of class \code{\link{docx}}.
}
\examples{
#START_TAG_TEST
# Create a new document
doc = docx( title = "title" )
print( doc )
#STOP_TAG_TEST
}
\seealso{
\code{\link{docx}}, \code{\link{print}}
}

