% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlexTable.R
\name{print.FlexTable}
\alias{print.FlexTable}
\title{Print FlexTables}
\usage{
\method{print}{FlexTable}(x, ...)
}
\arguments{
\item{x}{a \code{\link{FlexTable}} object}

\item{...}{further arguments, not used.}
}
\description{
print a \code{\link{FlexTable}} object.
If R session is interactive, the FlexTable is
rendered in an HTML page and loaded into a WWW browser.
}

