% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footnote.R
\name{print.Footnote}
\alias{print.Footnote}
\title{print a Footnote}
\usage{
\method{print}{Footnote}(x, ...)
}
\arguments{
\item{x}{a \code{\link{Footnote}} object}

\item{...}{further arguments, not used.}
}
\description{
print a Footnote
}

