% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot.R
\name{print.pot}
\alias{print.pot}
\title{Print pot objects}
\usage{
\method{print}{pot}(x, ...)
}
\arguments{
\item{x}{a \code{\link{pot}} object}

\item{...}{further arguments, not used.}
}
\description{
print a \code{\link{pot}} object.
Within RStudio, the pot is rendered in the viewer.
}

