\name{print.textProperties}
\alias{print.textProperties}
\title{print formating properties}
\usage{
  \method{print}{textProperties} (x, ...)
}
\arguments{
  \item{x}{an object of class \code{"textProperties"}}

  \item{...}{further arguments, not used.}
}
\description{
  print text formating properties (an object of class
  \code{"textProperties"}).
}
\examples{
print( textProperties (color="red", font.size = 12) )
}
\seealso{
  \code{\link{pptx}}, \code{\link{docx}}
}

