% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.plot.interactivity.R
\name{raphael_clicks}
\alias{raphael_clicks}
\title{clicks instructions to raphael device}
\usage{
raphael_clicks(ids, clicks)
}
\arguments{
\item{ids}{raphael elements identifiers. integer vector.}

\item{clicks}{javascript commands. character vector.}
}
\description{
send click instructions to raphael device. Internal use only.
}

