% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.plot.interactivity.R
\name{raphael_dbclicks}
\alias{raphael_dbclicks}
\title{tooltips instructions to raphael device}
\usage{
raphael_dbclicks(ids, dbclicks)
}
\arguments{
\item{ids}{raphael elements identifiers. integer vector.}

\item{dbclicks}{javascript commands. character vector.}
}
\description{
send tooltips instructions to raphael device. Internal use only.
}

