% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.plot.interactivity.R
\name{raphael_tooltips}
\alias{raphael_tooltips}
\title{tooltips instructions to raphael device}
\usage{
raphael_tooltips(ids, tooltips)
}
\arguments{
\item{ids}{raphael elements identifiers. integer vector.}

\item{tooltips}{tooltips. character vector.}
}
\description{
send tooltips instructions to raphael device. Internal use only.
}

