% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawingML.R
\name{registerRaphaelGraph}
\alias{registerRaphaelGraph}
\title{register Raphael plots}
\usage{
registerRaphaelGraph(plot_attributes, env)
}
\arguments{
\item{plot_attributes}{plot attributes}

\item{env}{environment}
}
\description{
register Raphael plots - internal use only
}

