% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderFlexTable.R
\name{renderFlexTable}
\alias{renderFlexTable}
\title{FlexTable output for shiny}
\usage{
renderFlexTable(expr, ..., env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a \code{FlexTable} object.}

\item{...}{not used}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Creates a reactive FlexTable that is suitable for
assigning to a \code{tableOutput} slot.
}
