% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FontMetric.R
\name{reporters_str_width}
\alias{reporters_str_width}
\title{Compute the width of a string}
\usage{
reporters_str_width(str, fontfamily, fontsize, fontface)
}
\arguments{
\item{str}{value}

\item{fontfamily}{font name}

\item{fontsize}{font size}

\item{fontface}{font face}
}
\description{
Compute the width of a string. Internal use.
}

