% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexTablePublicAPI.R
\name{setFlexTableWidths}
\alias{setFlexTableWidths}
\title{set columns widths of a FlexTable}
\usage{
setFlexTableWidths(object, widths)
}
\arguments{
\item{object}{a \code{FlexTable} object}

\item{widths}{a numeric vector specifying columns widths in inches.}
}
\description{
set columns widths of a FlexTable in inches.
}
\examples{
\donttest{
MyFTable <- FlexTable( data = iris[1:10, ] )
MyFTable <- setFlexTableWidths( MyFTable, widths = c(1,1,1,1,3))
}
}
\seealso{
\code{\link{FlexTable}}
}

