\name{setInteractiveElements}
\alias{setInteractiveElements}
\title{add interactivity on a plot}
\usage{
setInteractiveElements(fun, popup.labels, click.actions, dblclick.actions, ...)
}
\arguments{
  \item{fun}{plot function. See details.}

  \item{popup.labels}{labels to display when mouse is over
  the elements. A character vector. Length must be the same
  than the number of new elements generated by the plot
  function.}

  \item{click.actions}{events to run when mouse is clicking
  the elements. A character vector of javascript
  instructions. Length must be the same than the number of
  new elements generated by the plot function.}

  \item{dblclick.actions}{events to run when mouse is
  double-clicking the elements. A character vector of
  javascript instructions. Length must be the same than the
  number of new elements generated by the plot function.}

  \item{...}{arguments for \code{fun}.}
}
\description{
add interactivity on elements of a raphael plot. There are
three interactive features: popup text when mouse is over
an element, execute javascript instructions when clicking
the element and execute javascript instructions when
double-clicking the element.
}

