% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{slide.layouts}
\alias{slide.layouts}
\title{Get layout names of a document object}
\usage{
slide.layouts(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\description{
Get layout names that exist into a document
}
\details{
\code{slide.layouts} only works with pptx documents. See \code{\link{slide.layouts.pptx}} for examples.
}
\seealso{
\code{\link{pptx}}, \code{\link{slide.layouts.pptx}}, \code{\link{addSlide.pptx}}
}

