\name{spanCols}
\alias{spanCols}
\title{Span rows within columns}
\usage{
spanCols(object, i, from, to)
}
\arguments{
  \item{object}{a \code{FlexTable} object}

  \item{i}{vector integer index for columns.}

  \item{from}{first col to span (its content will be the
  visible one).}

  \item{to}{last col to span.}
}
\description{
Span columns within rows.
}
\seealso{
\code{\link{FlexTable}}
}

