\name{spanRows}
\alias{spanRows}
\title{Span rows within columns}
\usage{
spanRows(object, j, from, to)
}
\arguments{
  \item{object}{a \code{FlexTable} object}

  \item{j}{vector (integer index, col.names values or
  boolean vector) for columns.}

  \item{from}{first row to span (its content will be the
  visible one).}

  \item{to}{last row to span.}
}
\description{
Span rows within columns.
}
\seealso{
\code{\link{FlexTable}}
}

