% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.properties.R
\name{textBold}
\alias{textBold}
\title{shortcut for bold}
\usage{
textBold(...)
}
\arguments{
\item{...}{arguments passed to textProperties}
}
\description{
shortcut for bold textProperties()
}
\examples{
textBold()
}
\seealso{
\code{\link{textNormal}}
, \code{\link{textItalic}}
, \code{\link{textBoldItalic}}
, \code{\link{parRight}}
, \code{\link{parLeft}}
, \code{\link{parCenter}}
, \code{\link{parJustify}}
, \code{\link{borderDotted}}
, \code{\link{borderDashed}}
, \code{\link{borderNone}}
, \code{\link{borderSolid}}
}

