\name{textItalic}
\alias{textItalic}
\title{shortcut for italic}
\usage{
textItalic(...)
}
\arguments{
  \item{...}{arguments passed to textProperties}
}
\description{
shortcut for italic textProperties()
}
\examples{
textItalic()
}
\seealso{
\code{\link{textNormal}} , \code{\link{textBold}} ,
\code{\link{textBoldItalic}} , \code{\link{parRight}} ,
\code{\link{parLeft}} , \code{\link{parCenter}} ,
\code{\link{parJustify}} , \code{\link{borderDotted}} ,
\code{\link{borderDashed}} , \code{\link{borderNone}} ,
\code{\link{borderSolid}}
}

