\name{textNormal}
\alias{textNormal}
\title{shortcut for default textProperties}
\usage{
textNormal(...)
}
\arguments{
  \item{...}{arguments passed to textProperties}
}
\description{
shortcut for textProperties(...)
}
\examples{
textNormal()
}
\seealso{
\code{\link{textBold}} , \code{\link{textItalic}} ,
\code{\link{textBoldItalic}} , \code{\link{parRight}} ,
\code{\link{parLeft}} , \code{\link{parCenter}} ,
\code{\link{parJustify}} , \code{\link{borderDotted}} ,
\code{\link{borderDashed}} , \code{\link{borderNone}} ,
\code{\link{borderSolid}}
}

