\name{writeDoc}
\alias{writeDoc}
\title{Write a document object}
\usage{
  writeDoc(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
  a document object
}
\description{
  Write a document object into a file
}
\seealso{
  \code{\link{docx}}, \code{\link{pptx}},
  \code{\link{writeDoc.pptx}}
}

