% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addCodeBlock}
\alias{addCodeBlock}
\title{Add code block into a document object}
\usage{
addCodeBlock(doc, file, text, ...)
}
\arguments{
\item{doc}{document object}

\item{file}{script file. Not used if text is provided.}

\item{text}{character vector. The text to parse.
Not used if file is provided.}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a code block into a document object
}
\seealso{
\code{\link{addCodeBlock.bsdoc}}, \code{\link{addCodeBlock.docx}}
, \code{\link{addCodeBlock.pptx}}
}

