% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addColumnBreak}
\alias{addColumnBreak}
\title{Add a column break into a section}
\usage{
addColumnBreak(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a column break into a section
}
\details{
\code{addColumnBreak} only works with docx documents.

See \code{\link{addColumnBreak.docx}} for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{addColumnBreak.docx}}
}

