\name{addDate}
\alias{addDate}
\title{Insert a date into a document object}
\usage{
addDate(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Insert a column break
}
\details{
addDate only works for pptx documents. See
\code{\link{addSlide.pptx}}.

See \code{\link{addSlide.pptx}} for examples.
}
\seealso{
\code{\link{pptx}}, \code{\link{addSlide.pptx}}
}

