\name{addFlexTable}
\alias{addFlexTable}
\title{Insert a FlexTable into a document object}
\usage{
addFlexTable(doc, flextable, ...)
}
\arguments{
  \item{doc}{document object}

  \item{flextable}{the \code{FlexTable} object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Insert a FlexTable into a document object

FlexTable can be manipulated so that almost any formatting
can be specified. See \code{\link{FlexTable}} for more
details.
}
\details{
See \code{\link{addFlexTable.docx}} or
\code{\link{addFlexTable.pptx}} or
\code{\link{addFlexTable.bsdoc}} for examples.
}
\seealso{
\code{\link{FlexTable}}, \code{\link{addFlexTable.docx}} ,
\code{\link{addFlexTable.pptx}},
\code{\link{addFlexTable.bsdoc}}
}

