% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFooter.pptx.R
\name{addFooter.pptx}
\alias{addFooter.pptx}
\title{Insert a footer shape into a document pptx object}
\usage{
\method{addFooter}{pptx}(doc, value, ...)
}
\arguments{
\item{doc}{\code{\link{pptx}} object}

\item{value}{character value to add into the footer shape of the current slide.}

\item{...}{further arguments, not used.}
}
\value{
a document object
}
\description{
Insert a footer shape into the current slide of a \code{pptx} object.
}
\examples{
\donttest{
if( check_valid_java_version() ){
doc = pptx( )
doc = addSlide( doc, slide.layout = "Title and Content" )
doc = addFooter( doc, "Hi!" )
writeDoc( doc, file = "ex_footer.pptx" )
}
}
}
\seealso{
\code{\link{pptx}}, \code{\link{addDate.pptx}}
, \code{\link{addPageNumber.pptx}}
}
