\name{addIframe}
\alias{addIframe}
\title{Add an iframe into a document object}
\usage{
addIframe(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add an iframe into a document object
}
\seealso{
\code{\link{addIframe.bsdoc}}
}

