% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImage.pptx.R
\name{addImage.pptx}
\alias{addImage.pptx}
\title{Insert an external image into a pptx object}
\usage{
\method{addImage}{pptx}(doc, filename, offx, offy, width, height, ...)
}
\arguments{
\item{doc}{\code{\link{pptx}} object where external image has to be added}

\item{filename}{\code{"character"} value, complete filename of the external image}

\item{offx}{optional, x position of the shape (top left position of the bounding box) in inches. See details.}

\item{offy}{optional, y position of the shape (top left position of the bounding box) in inches See details.}

\item{width}{width of the shape in inches. See details.}

\item{height}{height of the shape in inches. See details.}

\item{...}{further arguments, not used.}
}
\value{
an object of class \code{\link{pptx}}.
}
\description{
Add an external image into a \code{\link{pptx}} object.
}
\details{
By default, image is added to the next free 'content' shape of the current slide.
See \code{\link{slide.layouts.pptx}} to view the slide layout.

Arguments \code{width} and \code{height} can be defined with functions
\code{png::readPNG}, \code{jpeg::readJPEG} or \code{bmp::read.bmp}.

If arguments offx and offy are missing, position is defined as
the position of the next available shape of the slide. This
dimensions can be defined in the layout of the PowerPoint template used to create
the \code{pptx} object.
}
\examples{
#START_TAG_TEST
doc.filename = "addImage_example.pptx"

# set default font size to 24
options( "ReporteRs-fontsize" = 24 )

doc = pptx( title = "title" )
img.file = file.path( Sys.getenv("R_HOME"), "doc", "html", "logo.jpg" )

if( file.exists( img.file ) && requireNamespace("jpeg", quietly = TRUE) ){

  dims <- attr( jpeg::readJPEG(img.file), "dim" )
  doc <- addSlide( doc, "Title and Content" )
  doc <- addTitle( doc, "Add images" )
  doc <- addImage(doc, img.file, width = dims[2]/72, height = dims[1]/72)
}
if( exists("win.metafile") ){
	win.metafile(filename = "image.wmf", width = 5, height = 5 )
	barplot( 1:6, col = 2:7)
	dev.off()

	doc = addSlide( doc, "Title and Content")
	doc = addImage(doc, "image.wmf", width = 5, height = 5 )
}

# Write the object
writeDoc( doc, file = doc.filename )
#STOP_TAG_TEST
}
\seealso{
\code{\link{pptx}}, \code{\link{addPlot.pptx}},
\code{\link{addImage}}
}

