\name{addImage}
\alias{addImage}
\title{Add an external image into a document object}
\usage{
  addImage(doc, filename, ...)
}
\arguments{
  \item{doc}{document object}

  \item{filename}{\code{"character"} value, complete
  filename of the external image}

  \item{...}{further arguments passed to other methods}
}
\value{
  a document object
}
\description{
  Add an external image into a document object
}
\seealso{
  \code{\link{docx}}, \code{\link{addImage.docx}} ,
  \code{\link{pptx}}, \code{\link{addImage.pptx}} ,
  \code{\link{html}}, \code{\link{addImage.html}}
}

