% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bsdoc.R
\name{addJavascript}
\alias{addJavascript}
\title{add javascript into a bsdoc object}
\usage{
addJavascript(doc, file, text)
}
\arguments{
\item{doc}{a \code{bsdoc} object.}

\item{file}{a javascript file. Not used if text is provided.}

\item{text}{character vector. The javascript text to parse.
Not used if file is provided.}
}
\value{
an object of class \code{\link{bsdoc}}.
}
\description{
add javascript into a \code{bsdoc} object.
}

