% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addMarkdown}
\alias{addMarkdown}
\title{Add a markdown text or file}
\usage{
addMarkdown(doc, file, text, ...)
}
\arguments{
\item{doc}{document object}

\item{file}{markdown file. Not used if text is provided.}

\item{text}{character vector. The markdown to parse.}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add markdown into a document object

The markdown definition used is John Gruber documented here:
\url{http://daringfireball.net/projects/markdown/syntax}.

Images are not available as \code{addImage} or \code{addPlot} is
available. Pandoc footnotes have been added (see
\url{http://johnmacfarlane.net/pandoc/README.html#footnotes}.
}
\seealso{
\code{\link{docx}}, \code{\link{addMarkdown.docx}}
, \code{\link{bsdoc}}, \code{\link{addMarkdown.bsdoc}}
, \code{\link{pptx}}, \code{\link{addMarkdown.pptx}}
}

