\name{addPage}
\alias{addPage}
\title{Add a page into a document object}
\usage{
addPage(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a page into a document object
}
\details{
\code{addPage} only works with html documents. See
\code{\link{addPage.html}} for examples.
}
\seealso{
\code{\link{html}}, \code{\link{addPage.html}}
}

