\name{addPageBreak}
\alias{addPageBreak}
\title{Add a page break into a document object}
\usage{
addPageBreak(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a page break into a document object
}
\seealso{
\code{\link{docx}}
}

