% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPageBreak.docx.R
\name{addPageBreak}
\alias{addPageBreak}
\alias{addPageBreak.docx}
\title{Add a page break into a document object}
\usage{
addPageBreak(doc, ...)

\method{addPageBreak}{docx}(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a page break into a document object
}
\details{
\code{addPageBreak} only works with docx documents.

See \code{\link{addPageBreak.docx}} for examples.
}
\examples{
\donttest{
doc = docx( title = "title" )
doc = addPageBreak( doc )
}
}
\seealso{
\code{\link{docx}}
}

