% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addPageNumber}
\alias{addPageNumber}
\title{Insert a page number into a document object}
\usage{
addPageNumber(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Insert a page number into a document object
}
\details{
\code{addPageNumber} only works with pptx documents.

See \code{\link{addPageNumber.pptx}} for examples.
}
\seealso{
\code{\link{pptx}}, \code{\link{addPageNumber.pptx}}
}

