% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawingML.R
\name{addPostCommand}
\alias{addPostCommand}
\title{add post plot commands}
\usage{
addPostCommand(labels, ids, env)
}
\arguments{
\item{labels}{labels}

\item{ids}{ids}

\item{env}{environment}
}
\description{
internal use only
}

