\name{addRScript}
\alias{addRScript}
\title{Add R script into a document object}
\usage{
addRScript(doc, rscript, ...)
}
\arguments{
  \item{doc}{document object}

  \item{rscript}{an object of class \code{RScript}}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add R script into a document object
}
\seealso{
\code{\link{addRScript.html}},
\code{\link{addRScript.docx}} ,
\code{\link{addRScript.pptx}}
}

