\name{addRScript}
\alias{addRScript}
\title{Add R script into a document object}
\usage{
  addRScript(doc, file, text, ...)
}
\arguments{
  \item{doc}{document object}

  \item{file}{R script file. Not used if text is provided.}

  \item{text}{character vector. The text to parse. Not used
  if file is provided.}

  \item{...}{further arguments passed to other methods}
}
\value{
  a document object
}
\description{
  Add R script into a document object
}
\seealso{
  \code{\link{html}}
}

