% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addSlide}
\alias{addSlide}
\title{Add a slide into a document object}
\usage{
addSlide(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a slide into a document object
}
\details{
\code{addSlide} only works with pptx documents. See \code{\link{addSlide.pptx}} for examples.
}
\seealso{
\code{\link{pptx}}, \code{\link{addSlide.pptx}}
}

