\name{addSubtitle}
\alias{addSubtitle}
\title{Add a subtitle shape into a document object}
\usage{
addSubtitle(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Add a subtitle shape into a document object
}
\seealso{
\code{\link{pptx}}
}

