% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{addTitle}
\alias{addTitle}
\title{Add a title into a document object}
\usage{
addTitle(doc, value, ...)
}
\arguments{
\item{doc}{document object}

\item{value}{\code{"character"} value to use as title text}

\item{...}{further arguments passed to or from other methods..}
}
\value{
a document object
}
\description{
Add a title into a document object
}
\details{
See \code{\link{addTitle.docx}} or \code{\link{addTitle.pptx}}
or \code{\link{addTitle.bsdoc}} for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{addTitle.docx}}, \code{\link{pptx}}
, \code{\link{addTitle.pptx}}, \code{\link{bsdoc}}, \code{\link{addTitle.bsdoc}}
}

