\name{addTOC}
\alias{addTOC}
\title{Add a table of contents into a document object}
\usage{
  addTOC(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
  a document object
}
\description{
  Add a table of contents into a document object
}
\seealso{
  \code{\link{docx}}, \code{\link{addTOC.docx}},
  \code{\link{styles.docx}}
}

