\name{borderNone}
\alias{borderNone}
\title{shortcut for no border}
\usage{
borderNone(...)
}
\arguments{
  \item{...}{arguments passed to borderProperties}
}
\description{
shortcut for no border borderProperties()
}
\examples{
borderNone()
}
\seealso{
\code{\link{textNormal}} , \code{\link{textBold}} ,
\code{\link{textItalic}} , \code{\link{textBoldItalic}} ,
\code{\link{parRight}} , \code{\link{parLeft}} ,
\code{\link{parCenter}} , \code{\link{parJustify}} ,
\code{\link{borderDotted}} , \code{\link{borderDashed}} ,
\code{\link{borderSolid}}
}

