% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borderProperties.R
\name{borderProperties}
\alias{borderProperties}
\alias{chprop.borderProperties}
\alias{print.borderProperties}
\alias{as.character.borderProperties}
\title{border properties object}
\usage{
borderProperties(color = "black", style = "solid", width = 1)

\method{chprop}{borderProperties}(object, color, style, width, ...)

\method{print}{borderProperties}(x, ...)

\method{as.character}{borderProperties}(x, ...)
}
\arguments{
\item{color}{border color - single character value (e.g. "#000000" or "black")}

\item{style}{border style - single character value : "none" or "solid" or "dotted" or "dashed"}

\item{width}{border width - an integer value : 0>= value}

\item{object}{\code{borderProperties} object to update}

\item{...}{further arguments - not used}

\item{x}{\code{borderProperties} object to print}
}
\description{
create a border properties object.
}
\details{
Get a modified version of a \code{borderProperties} with
\code{chprop}.
}
\examples{
borderProperties()
borderProperties(color="orange", style="solid", width=1)
borderProperties(color="gray", style="dotted", width=1)

# update borderProperties --------
x = borderProperties()
chprop(x, color="orange", style="dashed", width=1)
chprop(x, width=5)
}
\seealso{
\code{\link{alterFlexTable}}, \code{\link{setFlexTableBorders}},
\code{\link{shortcut_properties}}
}

