\name{chprop}
\alias{chprop}
\title{Change a formatting properties object}
\usage{
chprop(object, ...)
}
\arguments{
  \item{object}{formatting properties object}

  \item{...}{further arguments passed to other methods}
}
\value{
a formatting properties object
}
\description{
Change a formatting properties object
}
\seealso{
\code{\link{cellProperties}}, \code{\link{textProperties}}
, \code{\link{parProperties}}
}

