\name{declareTitlesStyles}
\alias{declareTitlesStyles}
\title{Set manually headers'styles of a document object}
\usage{
declareTitlesStyles(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Set manually titles'styles of a document object
}
\seealso{
\code{\link{docx}}, \code{\link{styles.docx}},
\code{\link{declareTitlesStyles.docx}}
}

