\name{deleteBookmark}
\alias{deleteBookmark}
\title{delete a bookmark into a docx object}
\usage{
deleteBookmark(doc, bookmark)
}
\arguments{
  \item{doc}{Object of class \code{"docx"}}

  \item{bookmark}{a character vector specifying bookmark id
  to delete}
}
\description{
delete a bookmark into a docx object
}
\seealso{
\code{\link{docx}}
}

