% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlexTable_theme.R
\name{light.table}
\alias{light.table}
\title{light FlexTable shortcut}
\usage{
light.table(dataset, add.rownames = FALSE, text.direction = "lrtb")
}
\arguments{
\item{dataset}{the data to use}

\item{add.rownames}{logical value - should the row.names be included in the table.}

\item{text.direction}{header cell text rotation - a single character value, expected
value is one of "lrtb", "tbrl", "btlr".}
}
\description{
get a simple FlexTable from a dataset
}
\examples{
\donttest{
ft <- light.table( iris)
}
}
\seealso{
\code{\link{FlexTable}}
}
