% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_extract.R
\name{list_bookmarks}
\alias{list_bookmarks}
\title{List Bookmarks from a Word Document}
\usage{
list_bookmarks(x, body = TRUE, header = TRUE, footer = TRUE)
}
\arguments{
\item{x}{a \code{docx} object}

\item{body}{specifies to scan document body}

\item{header}{specifies to scan document header}

\item{footer}{specifies to scan document footer}
}
\value{
a character vector
}
\description{
List all bookmarks available in a \code{docx} object.
}
\examples{
\donttest{
doc <- docx( title = "My example", template = file.path(
  system.file(package = "ReporteRs"), "templates/bookmark_example.docx") )
list_bookmarks( doc )
}
}
\seealso{
\code{\link{docx}}, \code{\link{text_extract}}
}

