% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot.R
\name{pot_img}
\alias{pot_img}
\title{Image to be concatenate with pot object}
\usage{
pot_img(filename, width, height)
}
\arguments{
\item{filename}{\code{"character"} value, complete filename of the external image}

\item{width}{image width in inches}

\item{height}{image height in inches}
}
\description{
Create an pot object that handle images.
}
