% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsdoc.R
\name{print.bsdoc}
\alias{print.bsdoc}
\title{Print method for \code{\link{bsdoc}} objects.}
\usage{
\method{print}{bsdoc}(x, ...)
}
\arguments{
\item{x}{a \code{\link{bsdoc}} object}

\item{...}{further arguments, not used.}
}
\description{
print a \code{\link{bsdoc}} object.
If R session is interactive, the document is
rendered in an HTML page and loaded into a WWW browser.
}
\examples{
#START_TAG_TEST
# Create a new document
doc = bsdoc( )
print( doc )
#STOP_TAG_TEST
}

