% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textProperties.R
\name{print.textProperties}
\alias{print.textProperties}
\title{print formatting properties}
\usage{
\method{print}{textProperties}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"textProperties"}}

\item{...}{further arguments, not used.}
}
\description{
print text formatting properties (an object of class \code{"textProperties"}).
}
\examples{
print( textProperties (color="red", font.size = 12) )
}
\seealso{
\code{\link{textProperties}}
}

