\name{registerRaphaelGraph}
\alias{registerRaphaelGraph}
\title{register Raphael plots}
\usage{
  registerRaphaelGraph(plot_attributes, env)
}
\arguments{
  \item{plot_attributes}{plot attributes}

  \item{env}{environment}
}
\description{
  register Raphael plots - internal use only
}

