% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.properties.R
\name{textNormal}
\alias{textNormal}
\alias{shortcut_properties}
\alias{textBold}
\alias{textItalic}
\alias{textBoldItalic}
\alias{parRight}
\alias{parLeft}
\alias{parCenter}
\alias{parJustify}
\alias{borderDotted}
\alias{borderDashed}
\alias{borderNone}
\alias{borderSolid}
\alias{cellBorderNone}
\alias{cellBorderBottom}
\alias{cellBorderTop}
\alias{cellBorderTB}
\title{shortcuts for formatting properties}
\usage{
textNormal(...)

textBold(...)

textItalic(...)

textBoldItalic(...)

parRight(...)

parLeft(...)

parCenter(...)

parJustify(...)

borderDotted(...)

borderDashed(...)

borderNone(...)

borderSolid(...)

cellBorderNone(...)

cellBorderBottom(...)

cellBorderTop(...)

cellBorderTB(...)
}
\arguments{
\item{...}{further arguments passed to original functions.}
}
\description{
Shortcuts for \code{textProperties}, \code{parProperties},
\code{borderProperties} and \code{cellProperties}.
}
\examples{
textNormal()
textBold()
textItalic()
textBoldItalic()
parRight()
parLeft()
parLeft()
parLeft()
borderDotted()
borderDashed()
borderNone()
borderSolid()
cellBorderNone()
cellBorderBottom()
cellBorderTop()
cellBorderTB()
}
