\name{slide.layouts}
\alias{slide.layouts}
\title{Get layout names of a document object}
\usage{
  slide.layouts(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\description{
  Get layout names that exist into a document
}
\seealso{
  \code{\link{pptx}}, \code{\link{slide.layouts.pptx}}
}

