% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_AllGeneric.R
\name{styles}
\alias{styles}
\title{Get styles names of a document object}
\usage{
styles(doc, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{further arguments passed to other methods}
}
\description{
Get styles names that exist into a document
}
\details{
\code{styles} only works with docx documents.

See \code{\link{styles.docx}} for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{styles.docx}}, \code{\link{addParagraph.docx}}
}

