\name{styles}
\alias{styles}
\title{Get styles names of a document object}
\usage{
styles(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\description{
Get styles names that exist into a document
}
\seealso{
\code{\link{docx}}, \code{\link{styles.docx}}
}

