% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FlexTable_theme.R
\name{vanilla.table}
\alias{vanilla.table}
\title{get a simple FlexTable from a dataset}
\usage{
vanilla.table(dataset, double.format = "\%0.3f", add.rownames = FALSE)
}
\arguments{
\item{dataset}{the data to use}

\item{double.format}{format string for \code{double} column to
format in the dataset. See argument \code{fmt} of \code{\link{sprintf}}.}

\item{add.rownames}{logical value - should the row.names be included in the table.}
}
\description{
get a simple FlexTable from a dataset
}
\examples{
#START_TAG_TEST
vanilla.table( iris)
#STOP_TAG_TEST
}
\seealso{
\code{\link{FlexTable}}
}

