% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FlexTable_theme.R
\name{vanilla.table}
\alias{vanilla.table}
\title{get a simple FlexTable from a dataset}
\usage{
vanilla.table(dataset, add.rownames = FALSE)
}
\arguments{
\item{dataset}{the data to use}

\item{add.rownames}{logical value - should the row.names be included in the table.}
}
\description{
get a simple FlexTable from a dataset
}
\examples{
vanilla.table( iris)
}
\seealso{
\code{\link{FlexTable}}
}

