% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDoc.docx.R
\name{writeDoc.docx}
\alias{writeDoc.docx}
\title{Write a docx object in a docx file}
\usage{
\method{writeDoc}{docx}(doc, file, ...)
}
\arguments{
\item{doc}{Object of class \code{\link{docx}} that has to be written.}

\item{file}{single character value, name of the file to write.}

\item{...}{further arguments, not used.}
}
\description{
Write the \code{\link{docx}} object in a '.docx' file.
}
\examples{
#START_TAG_TEST
doc.filename = "writeDoc_example.docx"

# set default font size to 10
options( "ReporteRs-fontsize" = 10 )

doc = docx( )

# Write the object
writeDoc( doc, file = doc.filename )
#STOP_TAG_TEST
}
\seealso{
\code{\link{docx}}, \code{\link{writeDoc}}
}

