% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDoc.R
\name{writeDoc}
\alias{writeDoc}
\alias{writeDoc.docx}
\alias{writeDoc.pptx}
\alias{writeDoc.bsdoc}
\title{Write a document object}
\usage{
writeDoc(doc, ...)

\method{writeDoc}{docx}(doc, file, ...)

\method{writeDoc}{pptx}(doc, file, ...)

\method{writeDoc}{bsdoc}(doc, file, ...)
}
\arguments{
\item{doc}{document object}

\item{...}{unused}

\item{file}{single character value, name of the html file to write.}
}
\description{
Write a document object into a file
}
\examples{
\donttest{
doc <- docx()
writeDoc( doc, "ex_write_doc.docx")

doc <- pptx()
doc <- addSlide(doc, "Title and Content")
writeDoc( doc, "ex_write_doc.pptx")
}
}
\seealso{
\code{\link{docx}}, \code{\link{pptx}}
}

