\name{writeDoc}
\alias{writeDoc}
\title{Write a document object}
\usage{
writeDoc(doc, ...)
}
\arguments{
  \item{doc}{document object}

  \item{...}{further arguments passed to other methods}
}
\value{
a document object
}
\description{
Write a document object into a file
}
\details{
See \code{\link{writeDoc.docx}} or
\code{\link{writeDoc.pptx}} or \code{\link{writeDoc.bsdoc}}
for examples.
}
\seealso{
\code{\link{docx}}, \code{\link{writeDoc.docx}} ,
\code{\link{pptx}}, \code{\link{writeDoc.pptx}} ,
\code{\link{bsdoc}}, \code{\link{writeDoc.bsdoc}}
}

